+++ libft/lst/ft_lstsize.c
@@ -6,15 +6,15 @@
 /*   By: omougel <omougel@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/20 14:03:38 by omougel           #+#    #+#             */
-/*   Updated: 2023/11/20 16:18:48 by omougel          ###   ########.fr       */
+/*   Updated: 2024/02/18 19:46:43 by omougel          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../include/libft.h"
 
-int	ft_lstsize(t_list *lst)
+size_t	ft_lstsize(t_list *lst)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (lst)
